% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_to_kleurring.R
\name{connect_to_kleurring}
\alias{connect_to_kleurring}
\title{Provide connection to the INBO kleurringdatabank}
\usage{
connect_to_kleurring()
}
\value{
conn ODBC connection to kleurring database.
}
\description{
Connect to the INBO kleurringdatabank (the so called
  meeuwen database). Be sure to have the needed rights by INBO databeheer
  team. The driver `SQL Server Native Client 11.0` is used.
}
