% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_to_kleurring.R
\docType{data}
\name{driver}
\alias{driver}
\title{kleurring database informations}
\format{An object of class \code{character} of length 1.}
\usage{
driver
}
\description{
kleurring database informations
}
\keyword{datasets}
