% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birds.R
\name{get_birds}
\alias{get_birds}
\title{Get birds data}
\usage{
get_birds(conn)
}
\arguments{
\item{conn}{A valid connection to INBO kleurring database.}
}
\value{
A tibble data frame.
}
\description{
Get birds (ring) data.
}
\examples{
\dontrun{
conn <- connect_to_kleurring()

# all birds
birds <- sovon::get_birds(conn)
}
}
