% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Get observations data}
\usage{
get_observations(conn)
}
\arguments{
\item{conn}{A valid connection to INBO kleurring database.}
}
\value{
A tibble data frame.
}
\description{
Get observations data.
}
\examples{
\dontrun{
conn <- connect_to_kleurring()

# all observations
observations <- sovon::get_observations(conn)
}
}
