% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get users data}
\usage{
get_users(conn)
}
\arguments{
\item{conn}{A valid connection to INBO kleurring database.}
}
\value{
A tibble data frame.
}
\description{
Get users data.
}
\examples{
\dontrun{
conn <- connect_to_kleurring()

# all users
users <- sovon::get_users(conn)
}
}
