% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_birds.R
\docType{data}
\name{sovon_birds_fields}
\alias{sovon_birds_fields}
\title{birds fields requested by SOVON}
\format{An object of class \code{character} of length 12.}
\usage{
sovon_birds_fields
}
\description{
birds fields requested by SOVON
}
\keyword{datasets}
