% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_users.R
\docType{data}
\name{sovon_user_fields}
\alias{sovon_user_fields}
\title{user fields requested by SOVON}
\format{An object of class \code{character} of length 11.}
\usage{
sovon_user_fields
}
\description{
user fields requested by SOVON
}
\keyword{datasets}
