% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_birds.R
\name{transform_birds}
\alias{transform_birds}
\title{Transform bird data to SOVON format}
\usage{
transform_birds(birds)
}
\arguments{
\item{birds}{tibble data frame.}
}
\value{
A tibble data frame.
}
\description{
SOVON requests bird data in a specific format. This function tranforms the
information extracted from INBO kleurring database to such format.
}
