% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_users.R
\name{transform_users}
\alias{transform_users}
\title{Transform user data to SOVON format}
\usage{
transform_users(users)
}
\arguments{
\item{users}{tibble data frame.}
}
\value{
A tibble data frame.
}
\description{
SOVON requests user data in a specific format. This function tranforms the
information extracted from INBO kleurring database to such format.
}
